split comvis qss
